@@ -1,1 +1,1 @@
-ruby-2.2.0
+ruby-2.4.0
